Fall back to an "adequate" split point
If a line is so long that after splitting one of the lines would still be too long SE falls back to a roughly central split. However, it's rarely a good choice as it will split mid-sentence, mid-name etc.

There usually _is_ a natural split point in very long lines (probably more than one, that's why it's so long) so if we can't find a "good" split point look for an "OK" one instead.

This usually seems to result in a pretty good split with one or both of the split lines being in good shape for splitting again.